refactor: private registry
